Implemented internally usable packet building. No bandwidth considerations yet. Deleted defunct tests files

+++ src/common/com_debug.h
@@ -4,4 +4,4 @@
 
 void COM_PrintBits(i32 val, u8 printNewLine);
 void COM_PrintBytes(u8* bytes, i32 numBytes, i32 bytesPerRow);
-void COM_PrintBytesHex(u8* bytes, u16 numBytes, i32 bytesPerRow);
+void COM_PrintBytesHex(u8* bytes, i32 numBytes, i32 bytesPerRow);
